apps/mcp-offers/pyproject.toml
[project]
name = "mcp-offers"
version = "0.1.0"
description = "MCP server for ACP offers search and discovery"
authors = [
    {name = "ACP Demo Team", email = "demo@example.com"}
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
